+++ lib/src/app_strings.dart
@@ -1,0 +1,2 @@
+// Community strings
+const String groupText = 'Group';